diplomat-gen

+++ ffi/capi/bindings/c/ICU4XDateTime.h
@@ -50,7 +50,7 @@ uint8_t ICU4XDateTime_second(const ICU4XDateTime* self);
 
 uint32_t ICU4XDateTime_nanosecond(const ICU4XDateTime* self);
 
-uint32_t ICU4XDateTime_day_of_year(const ICU4XDateTime* self);
+uint16_t ICU4XDateTime_day_of_year(const ICU4XDateTime* self);
 
 uint32_t ICU4XDateTime_day_of_month(const ICU4XDateTime* self);
