+++ ffi/capi/bindings/cpp/ICU4XDate.hpp
@@ -70,7 +70,7 @@ class ICU4XDate {
    * 
    * See the [Rust documentation for `day_of_year_info`](https://docs.rs/icu/latest/icu/calendar/struct.Date.html#method.day_of_year_info) for more information.
    */
-  uint32_t day_of_year() const;
+  uint16_t day_of_year() const;
 
   /**
    * Returns the 1-indexed day in the month for this date
@@ -223,7 +223,7 @@ inline ICU4XDate ICU4XDate::to_calendar(const ICU4XCalendar& calendar) const {
 inline ICU4XIsoDate ICU4XDate::to_iso() const {
   return ICU4XIsoDate(capi::ICU4XDate_to_iso(this->inner.get()));
 }
-inline uint32_t ICU4XDate::day_of_year() const {
+inline uint16_t ICU4XDate::day_of_year() const {
   return capi::ICU4XDate_day_of_year(this->inner.get());
 }
 inline uint32_t ICU4XDate::day_of_month() const {
