+++ ffi/capi/bindings/cpp/ICU4XDateTime.hpp
@@ -121,7 +121,7 @@ class ICU4XDateTime {
    * 
    * See the [Rust documentation for `day_of_year_info`](https://docs.rs/icu/latest/icu/calendar/struct.Date.html#method.day_of_year_info) for more information.
    */
-  uint32_t day_of_year() const;
+  uint16_t day_of_year() const;
 
   /**
    * Returns the 1-indexed day in the month for this date
@@ -293,7 +293,7 @@ inline uint8_t ICU4XDateTime::second() const {
 inline uint32_t ICU4XDateTime::nanosecond() const {
   return capi::ICU4XDateTime_nanosecond(this->inner.get());
 }
-inline uint32_t ICU4XDateTime::day_of_year() const {
+inline uint16_t ICU4XDateTime::day_of_year() const {
   return capi::ICU4XDateTime_day_of_year(this->inner.get());
 }
 inline uint32_t ICU4XDateTime::day_of_month() const {
