+++ ffi/capi/bindings/cpp/ICU4XIsoDate.hpp
@@ -67,7 +67,7 @@ class ICU4XIsoDate {
    * 
    * See the [Rust documentation for `day_of_year_info`](https://docs.rs/icu/latest/icu/calendar/struct.Date.html#method.day_of_year_info) for more information.
    */
-  uint32_t day_of_year() const;
+  uint16_t day_of_year() const;
 
   /**
    * Returns the 1-indexed day in the month for this date
@@ -175,7 +175,7 @@ inline ICU4XDate ICU4XIsoDate::to_calendar(const ICU4XCalendar& calendar) const 
 inline ICU4XDate ICU4XIsoDate::to_any() const {
   return ICU4XDate(capi::ICU4XIsoDate_to_any(this->inner.get()));
 }
-inline uint32_t ICU4XIsoDate::day_of_year() const {
+inline uint16_t ICU4XIsoDate::day_of_year() const {
   return capi::ICU4XIsoDate_day_of_year(this->inner.get());
 }
 inline uint32_t ICU4XIsoDate::day_of_month() const {
