+++ ffi/capi/bindings/js/ICU4XDateTime.d.ts
@@ -113,7 +113,7 @@ export class ICU4XDateTime {
 
    * See the {@link https://docs.rs/icu/latest/icu/calendar/struct.Date.html#method.day_of_year_info Rust documentation for `day_of_year_info`} for more information.
    */
-  day_of_year(): u32;
+  day_of_year(): u16;
 
   /**
